a2-transcoder-main/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic
python-multipart
requests
PyJWT
boto3
botocore